+++ src/python/wmagentscripts.egg-info/PKG-INFO
@@ -1,10 +1,12 @@
-Metadata-Version: 1.0
+Metadata-Version: 2.1
 Name: wmagentscripts
 Version: 0.0.1
 Summary: WMAgentScripts
 Home-page: UNKNOWN
 Author: Hasan Ozturk
-Author-email: UNKNOWN
 License: UNKNOWN
-Description: UNKNOWN
 Platform: UNKNOWN
+Requires-Python: >=3.6
+
+UNKNOWN
+
